Done!
The only thing I skipped was the usage of other queries, but I promise I'm going continue working that directly from mongoDB. Also It take time for me to get the whole concept of MODELS and 1-1, 1-many, many-many but now I can see it everywhere! wahoo! let's go for the other classes